+++ app/controllers/datasets_controller.rb
@@ -73,6 +73,7 @@ class DatasetsController < ApplicationController
     end
     scope = scope.where(polysomnography: true) if params[:data] == "polysomnography"
     scope = scope.where(actigraphy: true) if params[:data] == "actigraphy"
+    scope = scope.where(questionnaires_only: true) if params[:data] == "questionnaires-only"
     @datasets = scope_order(scope).page(params[:page]).per(25)
   end