+++ db/migrate/20230325184050_add_questionnaires_only_to_datasets.rb
@@ -1,0 +1,6 @@
+class AddQuestionnairesOnlyToDatasets < ActiveRecord::Migration[6.1]
+  def change
+    add_column :datasets, :questionnaires_only, :boolean, null: false, default: false
+    add_index :datasets, :questionnaires_only
+  end
+end