+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **General Changes**
  - Added a new navy theme for the website
  - Added new logo to landing page and corresponding favicon
+- **Dataset Changes**
+  - Added "Questionnaires Only" option for datasets
 
 ### Bug Fixes
 - Adjust menu navbar to prevent elements from moving when hovering on windows